+++ src/onedrivesdk/model/share.py
@@ -88,3 +88,4 @@ class Share(OneDriveObjectBase):
         else:
             return None
 
+from ..model.items_collection_page import ItemsCollectionPage
