+++ src/python2/request/drive_request.py
@@ -1,24 +1,6 @@
 # -*- coding: utf-8 -*- 
 '''
-# Copyright (c) 2015 Microsoft Corporation
-# 
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-# 
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
-# 
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-# THE SOFTWARE.
+# Copyright (c) Microsoft Corporation.  All Rights Reserved.  Licensed under the MIT License.  See License in the project root for license information.
 # 
 #  This file was generated and any changes will be overwritten.
 '''
@@ -79,15 +61,15 @@ class DriveRequest(RequestBase):
 
     def _initialize_collection_properties(self, value):
         if value and value._prop_dict:
-            if value.items and value.items._prop_dict:
+            if value.items:
                 if "items@odata.nextLink" in value._prop_dict:
                     next_page_link = value._prop_dict["items@odata.nextLink"]
-                    value.items._init_next_page_request(next_page_link, self._client, None)
-            if value.shared and value.shared._prop_dict:
+                    value.items._next_page_link = next_page_link
+            if value.shared:
                 if "shared@odata.nextLink" in value._prop_dict:
                     next_page_link = value._prop_dict["shared@odata.nextLink"]
-                    value.shared._init_next_page_request(next_page_link, self._client, None)
-            if value.special and value.special._prop_dict:
+                    value.shared._next_page_link = next_page_link
+            if value.special:
                 if "special@odata.nextLink" in value._prop_dict:
                     next_page_link = value._prop_dict["special@odata.nextLink"]
-                    value.special._init_next_page_request(next_page_link, self._client, None)
+                    value.special._next_page_link = next_page_link
