+++ src/python2/request/thumbnails_collection.py
@@ -1,32 +1,14 @@
 # -*- coding: utf-8 -*- 
 '''
-# Copyright (c) 2015 Microsoft Corporation
-# 
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-# 
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
-# 
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-# THE SOFTWARE.
+# Copyright (c) Microsoft Corporation.  All Rights Reserved.  Licensed under the MIT License.  See License in the project root for license information.
 # 
 #  This file was generated and any changes will be overwritten.
 '''
 
 from __future__ import unicode_literals
-from ..collection_base import CollectionRequestBase, CollectionResponseBase, CollectionPageBase
+from ..collection_base import CollectionRequestBase, CollectionResponseBase
 from ..request_builder_base import RequestBuilderBase
-from ..model.thumbnail_set import ThumbnailSet
+from ..model.thumbnails_collection_page import ThumbnailsCollectionPage
 import json
 
 class ThumbnailsCollectionRequest(CollectionRequestBase):
@@ -55,6 +37,19 @@ class ThumbnailsCollectionRequest(CollectionRequestBase):
         collection_response = ThumbnailsCollectionResponse(json.loads(self.send().content))
         return self._page_from_response(collection_response)
 
+
+    @staticmethod
+    def get_next_page_request(collection_page, client, options):
+        """Gets the ThumbnailsCollectionRequest for the next page. Returns None if there is no next page
+
+        Yields: 
+            :class:`ThumbnailsCollectionRequest<onedrivesdk.request.thumbnails_collection.ThumbnailsCollectionRequest>`:
+                The ThumbnailsCollectionRequest
+        """
+        if collection_page._next_page_link:
+            return ThumbnailsCollectionRequest(collection_page._next_page_link, client, options)
+        else:
+            return None
 
 class ThumbnailsCollectionRequestBuilder(RequestBuilderBase):
 
@@ -119,44 +114,6 @@ class ThumbnailsCollectionResponse(CollectionResponseBase):
             self._collection_page = ThumbnailsCollectionPage(self._prop_dict["value"])
 
         return self._collection_page
-
-
-class ThumbnailsCollectionPage(CollectionPageBase):
-
-    def __getitem__(self, index):
-        """Get the ThumbnailSet at the index specified
-        
-        Args:
-            index (int): The index of the item to get from the ThumbnailsCollectionPage
-
-        Returns:
-            :class:`ThumbnailSet<onedrivesdk.model.thumbnail_set.ThumbnailSet>`:
-                The ThumbnailSet at the index
-        """
-        return ThumbnailSet(self._prop_list[index])
-
-    def thumbnails(self):
-        """Get a generator of ThumbnailSet within the ThumbnailsCollectionPage
-        
-        Yields:
-            :class:`ThumbnailSet<onedrivesdk.model.thumbnail_set.ThumbnailSet>`:
-                The next ThumbnailSet in the collection
-        """
-        for item in self._prop_list:
-            yield ThumbnailSet(item)
-
-    def _init_next_page_request(self, next_page_link, client, options):
-        """Initialize the next page request for the ThumbnailsCollectionPage
-        
-        Args:
-            next_page_link (str): The URL for the next page request
-                to be sent to
-            client (:class:`OneDriveClient<onedrivesdk.model.one_drive_client.OneDriveClient>`:
-                The client to be used for the request
-            options (list of :class:`Option<onedrivesdk.options.Option>`:
-                A list of options
-        """
-        self._next_page_request = ThumbnailsCollectionRequest(next_page_link, client, options)
 
 
 from ..request.thumbnail_set_request_builder import ThumbnailSetRequestBuilder
